make sure git-apply-compatible prefix is present
... that was a lot harder than it should have been x.x

+++ scripts/dev/format_pull_request.php
@@ -1,8 +1,9 @@
 <?php
+
 declare(strict_types=1);
 error_reporting(E_ALL);
-set_error_handler(function ($errno, $errstr, $errfile, $errline){
-    if(error_reporting() & $errno) {
+set_error_handler(function ($errno, $errstr, $errfile, $errline) {
+    if (error_reporting() & $errno) {
         throw new ErrorException($errstr, 0, $errno, $errfile, $errline);
     }
 });
@@ -21,10 +22,10 @@ if ($argc !== 2) {
     exit(1);
 }
 chdir(__DIR__ . '/../../'); // php-src root
-if(!$force){
+if (!$force) {
     $gitStatus = null;
     system('git diff --quiet --exit-code', $gitStatus);
-    if($gitStatus !== 0){
+    if ($gitStatus !== 0) {
         fwrite(STDERR, "Error: Working directory is not clean according to 'git diff'\n");
         fwrite(STDERR, "commit your stuff -or- run 'git reset --hard' and try again\n");
         exit(1);
@@ -37,9 +38,20 @@ $diffFileHandle = tmpfile();
 fwrite($diffFileHandle, $diff);
 $diffFilePath = stream_get_meta_data($diffFileHandle)['uri'];
 passthru("git apply " . escapeshellarg($diffFilePath), $ret);
-if($ret !== 0){
+if ($ret !== 0) {
     fwrite(STDERR, "Error: git apply failed with exit code {$ret}\n");
-    passthru("git reset --hard");
+    if (!$force) {
+        passthru('git reset --hard');
+    }
+    exit(1);
+}
+// commit
+passthru('git commit -a --message=clangformat --quiet', $ret);
+if ($ret !== 0) {
+    fwrite(STDERR, "Error: git commit failed with exit code {$ret}\n");
+    if (!$force) {
+        passthru('git reset --hard');
+    }
     exit(1);
 }
 $clangFormatOptions = array(
@@ -48,21 +60,27 @@ $clangFormatOptions = array(
     'BasedOnStyle' => 'WebKit',
     'UseTab' => 'ForIndentation',
     // clang-format's support for tabs isn't great, and we have to say "4 spaces" to get 1 tab
-    'IndentWidth' => 4, 
+    'IndentWidth' => 4,
     'TabWidth' => 4,
 );
-
 $cmd = array(
-    'git diff -U0 --no-color --relative  |',
-    'clang-format-diff -p1 -style=' . escapeshellarg(json_encode($clangFormatOptions)),
+    'git diff -U0 --no-color --relative HEAD^ |',
+    'clang-format-diff -i -p1 -style=' . escapeshellarg(json_encode($clangFormatOptions)),
     '-iregex=' . escapeshellarg('.*\.(c|cc|cpp|c\+\+|cxx|h|hh|hpp|h\+\+|h)$'),
 );
 $cmd = implode(' ', $cmd);
 passthru($cmd, $ret);
-if($ret !== 0){
+if ($ret !== 0) {
     fwrite(STDERR, "Error: clang-format-diff failed with exit code {$ret}\n");
     fwrite(STDERR, "cmd:\n{$cmd}\n");
-    passthru("git reset --hard");
+    if (!$force) {
+        // remove last commit and reset
+        passthru('git reset --hard HEAD^ --quiet');
+    }
     exit(1);
 }
-shell_exec("git reset --hard");
+passthru('git diff');
+if (!$force) {
+    // remove last commit and reset
+    passthru('git reset --hard HEAD^ --quiet');
+}
